new structure

+++ terraform/hub/destroy.sh
@@ -20,9 +20,13 @@ pushd ${SCRIPTDIR}
 # Delete the Ingress/SVC before removing the addons
 TMPFILE=$(mktemp)
 terraform output -raw configure_kubectl > "$TMPFILE"
-source "$TMPFILE"
+# check if TMPFILE contains the string "No outputs found"
+if [[ ! $(cat $TMPFILE) == *"No outputs found"* ]]; then
+  echo "No outputs found, skipping kubectl delete"
+  source "$TMPFILE"
+  kubectl delete svc -n argocd argo-cd-argocd-server
+fi
 
-kubectl delete svc -n argocd argo-cd-argocd-server
 
 terraform destroy -target="module.gitops_bridge_bootstrap" -auto-approve
 terraform destroy -target="module.eks_blueprints_addons" -auto-approve
